+++ docs/dev/table/tableApi.md
@@ -1785,12 +1785,25 @@ The `OverWindow` defines a range of rows over which aggregates are computed. `Ov
 
 {% top %}
 
-### Map
-
-Map performs a map operation with an user-defined scalar function or built-in scalar function. The output will be flattened if the output type is a composite type.
+### Row-based Operations
 
 <div class="codetabs" markdown="1">
 <div data-lang="java" markdown="1">
+<table class="table table-bordered">
+  <thead>
+    <tr>
+      <th class="text-left" style="width: 20%">Operators</th>
+      <th class="text-center">Description</th>
+    </tr>
+  </thead>
+  <tbody>
+    <tr>
+      <td>
+        <strong>Map</strong><br>
+        <span class="label label-primary">Batch</span> <span class="label label-primary">Streaming</span>
+      </td>
+      <td>
+        <p>Map performs a map operation with a user-defined scalar function or built-in scalar function. The output will be flattened if the output type is a composite type.</p>
 {% highlight java %}
 ScalarFunction func = new MyMapFunction();
 tableEnv.registerFunction("func", func);
@@ -1798,15 +1811,40 @@ tableEnv.registerFunction("func", func);
 Table table = input
   .map(func("c")).as("a, b")
 {% endhighlight %}
+      </td>
+    </tr>
+
+  </tbody>
+</table>
 </div>
 
 <div data-lang="scala" markdown="1">
+<table class="table table-bordered">
+  <thead>
+    <tr>
+      <th class="text-left" style="width: 20%">Operators</th>
+      <th class="text-center">Description</th>
+    </tr>
+  </thead>
+  <tbody>
+    <tr>
+      <td>
+        <strong>Map</strong><br>
+        <span class="label label-primary">Batch</span> <span class="label label-primary">Streaming</span>
+      </td>
+      <td>
+        <p>Map performs a map operation with a user-defined scalar function or built-in scalar function. The output will be flattened if the output type is a composite type.</p>
 {% highlight scala %}
 val func: ScalarFunction = new MyMapFunction()
 
 val table = input
   .map(func('c)).as('a, 'b)
 {% endhighlight %}
+      </td>
+    </tr>
+
+  </tbody>
+</table>
 </div>
 </div>
